Model is refactored
