+++ Irrigation/include/Model/IrrigationDocument.h
@@ -1,32 +1,26 @@
 #pragma once
-
-#include <array>
-#include <mutex>
-
 #include "DocumentView/Document.h"
-#include "Logic/Containers.h"
-#include "Logic/Program.h"
-#include "Logic/Zones.h"
+#include "Logic/ProgramContainer.h"
+#include "Logic/ZoneHandler.h"
 
 
 class IrrigationDocument : public Document {
 
-	std::unique_ptr<ProgramContainer> programs;
-	Zones zones;
+	ProgramContainer programs;
+	ZoneHandler zones;
 
 public:
 
 	IrrigationDocument();
 	virtual ~IrrigationDocument();
 
-
 	// Program
-	const ProgramContainer& getPrograms() const { return *programs; }
-	ProgramContainer& getPrograms() { return *programs; }
+	const ProgramContainer& getPrograms() const { return programs; }
+	ProgramContainer& getPrograms() { return programs; }
 
 	// Zones
-	const Zones& getZones() const { return zones; }
-	Zones& getZones() { return zones; }
+	const ZoneHandler& getZones() const { return zones; }
+	ZoneHandler& getZones() { return zones; }
 
 	virtual void load();
 };
