+++ Irrigation/src/Model/IrrigationDocument.cpp
@@ -1,15 +1,13 @@
-#include "common.h"
 #include "IrrigationDocument.h"
-
-#include "Logic/ContainerFactories.h"
 #include "Logic/Program.h"
+#include "Logic/RunTime.h"
+#include "Logic/RunTimeContainer.h"
+#include "Logic/StartTime.h"
+#include "Logic/StartTimeContainer.h"
 
 
 
-IrrigationDocument::IrrigationDocument() :
-	programs(ProgramContainerFactoryHolder::create()),
-	zones(MASTER_VALVE_ID, ZONE_VALVE_IDS)
-{
+IrrigationDocument::IrrigationDocument()  {
 }
 
 IrrigationDocument::~IrrigationDocument() {
@@ -18,20 +16,23 @@ IrrigationDocument::~IrrigationDocument() {
 void IrrigationDocument::load() {
 	Program* program;
 
-	program = new Program("fulocsolas");
-	program->getRunTimes().modify(0, 25);
-	program->getRunTimes().modify(1, 35);
-	program->getRunTimes().modify(2, 25);
-	program->getStartTimes().insert(StartTime(4, 0));
+	program = new Program();
+	program->setName("fulocsolas");
+	program->getRunTimes().at(0)->setValue(25);
+	program->getRunTimes().at(1)->setValue(35);
+	program->getRunTimes().at(2)->setValue(25);
+	program->getStartTimes().insert(std::make_pair(0, new StartTime(4, 0, 0)));
 	getPrograms().insert(program);
 
-	program = new Program("buxus");
-	program->getRunTimes().modify(3, 20);
-	program->getStartTimes().insert(StartTime(6, 0));
+	program = new Program();
+	program->setName("buxus");
+	program->getRunTimes().at(3)->setValue(20);
+	program->getStartTimes().insert(std::make_pair(1, new StartTime(6, 0, 0)));
 	getPrograms().insert(program);
 
-	program = new Program("paradicsom");
-	program->getRunTimes().modify(4, 3);
-	program->getStartTimes().insert(StartTime(8, 0));
+	program = new Program();
+	program->setName("paradicsom");
+	program->getRunTimes().at(4)->setValue(3);
+	program->getStartTimes().insert(std::make_pair(2, new StartTime(8, 0, 0)));
 	getPrograms().insert(program);
 }
